+++ src/cyclops/CyclicCoordinateDescent.cpp
@@ -535,7 +535,7 @@ void CyclicCoordinateDescent::setCensorWeights(double* dWeights) {
 }
 
 double CyclicCoordinateDescent::getLogPrior(void) {
-	return jointPrior->logDensity(hBeta);
+	return jointPrior->logDensity(hBeta, *this);
 }
 
 double CyclicCoordinateDescent::getObjectiveFunction(int convergenceType) {
@@ -844,6 +844,7 @@ bool CyclicCoordinateDescent::performCheckConvergence(int convergenceType,
     bool done = false;
     double conv;
     bool illconditioned = false;
+    bool poorBlr = false;
     if (convergenceType < ZHANG_OLES) {
         double thisObjFunc = getObjectiveFunction(convergenceType);
         if (thisObjFunc != thisObjFunc) {
@@ -857,6 +858,16 @@ bool CyclicCoordinateDescent::performCheckConvergence(int convergenceType,
             illconditioned = true;
         } else {
             conv = computeConvergenceCriterion(thisObjFunc, *lastObjFunc);
+            if (conv == 0.0 && iteration == 1 && convergenceType != ONE_STEP &&
+                !std::all_of(std::begin(fixBeta), std::end(fixBeta), [](bool fixed) { return fixed; })
+            ) {
+                std::ostringstream stream;
+                stream << "\nWarning: BLR gradient is ill-conditioned\n"
+                       << "Enforcing convergence!";
+                logger->writeLine(stream);
+                illconditioned = true;
+                poorBlr = true;
+            }
         }
         *lastObjFunc = thisObjFunc;
     } else { // ZHANG_OLES
@@ -880,7 +891,9 @@ bool CyclicCoordinateDescent::performCheckConvergence(int convergenceType,
     }
 
     if (epsilon > 0 && conv < epsilon) {
-        if (illconditioned) {
+        if (poorBlr) {
+            lastReturnFlag = POOR_BLR_STEP;
+        } else if (illconditioned) {
             lastReturnFlag = ILLCONDITIONED;
         } else {
             if (noiseLevel > SILENT) {
@@ -1057,148 +1070,148 @@ void CyclicCoordinateDescent::findMode(
 	qnQ = 0;
 
 	if (qnQ > 0) { // Use quasi-Newton
-	    using namespace Eigen;
-
-	    Eigen::MatrixXd secantsU(J, qnQ);
-	    Eigen::MatrixXd secantsV(J, qnQ);
-
-	    VectorXd x(J);
-
-	    // Fill initial secants
-	    int countU = 0;
-	    int countV = 0;
-
-	    for (int q = 0; q < qnQ; ++q) {
-	        x = Map<const VectorXd>(hBeta.data(), J); // Make copy
-
-	        cycle();
-	        done = check();
-	        if (done) break;
-
-	        if (countU == 0) { // First time through
-	            secantsU.col(countU) = Map<const VectorXd>(hBeta.data(), J) - x;
-	            ++countU;
-	        } else if (countU < qnQ - 1) { // Middle case
-	            secantsU.col(countU) = Map<const VectorXd>(hBeta.data(), J) - x;
-	            secantsV.col(countV) = secantsU.col(countU);
-	            ++countU;
-	            ++countV;
-	        } else { // Last time through
-	            secantsV.col(countV) = Map<const VectorXd>(hBeta.data(), J) - x;
-	            ++countV;
-	        }
-	    }
-
-	    // 	    std::cerr << secantsU << std::endl << std::endl;
-	    // 	    std::cerr << secantsV << std::endl;
-
-	    int newestSecant = qnQ - 1;
-	    int previousSecant = newestSecant - 1;
-
-	    while (!done) {
-
-	        // 2 cycles for each QN step
-	        x = Map<const VectorXd>(hBeta.data(), J); // Make copy
-	        cycle();
-	        // 	        done = check();
-	        // 	        if (done) {
-	        // 	            std::cerr << "break A" << std::endl;
-	        // 	            break;
-	        // 	        }
-
-	        secantsU.col(newestSecant) = Map<const VectorXd>(hBeta.data(), J) - x;
-
-	        VectorXd Fx = Map<const VectorXd>(hBeta.data(), J); // TODO Can remove?
-
-	        x = Map<const VectorXd>(hBeta.data(), J);
-	        cycle();
-	        // 	        done = check();
-	        // 	        if (done) {
-	        // 	            std::cerr << "break B" << std::endl;
-	        // 	            break;
-	        // 	        }
-
-	        secantsV.col(newestSecant) = Map<const VectorXd>(hBeta.data(), J) - x;
-
-	        // Do QN step here
-
-	        //             MatrixXd UtU = secantsU.transpose() * secantsU;
-	        //             MatrixXd UtV = secantsU.transpose() * secantsV;
-	        //             MatrixXd M = UtU - UtV;
-
-	        auto M = secantsU.transpose() * (secantsU - secantsV);
-
-	        auto Minv = M.inverse();
-
-	        auto A = secantsU.transpose() * secantsU.col(newestSecant);
-	        auto B = Minv * A;
-	        auto C = secantsV * B;
-
-	        // MatrixXd A = secantsV * Minv;
-	        // MatrixXd B = A * secantsU.transpose();
-	        // MatrixXd C = B * secantsV.col(newestSecant); // - (x - Fx)
-	        // MatrixXd C = B * secantsU.col(newestSecant); // TODO Can remove?
-
-	        //     VectorXd xqn = Map<const VectorXd>(hBeta.data(), J) + C;
-	        VectorXd xqn = Fx + C; // TODO Can remove?
-
-	        // Save CCD solution
-	        x = Map<const VectorXd>(hBeta.data(), J);
-	        double ccdObjective = getLogLikelihood() + getLogPrior();
-	        // double savedLastObjFunc = lastObjFunc;
-
-	        Map<VectorXd>(hBeta.data(), J) = xqn; // Set QN solution
-	        //             for (int j = 0; j < J; ++j) {
-	        //                 if (sign(hBeta[j]) == xqn(j)) {
-	        //                     hBeta[j] = xqn(j);
-	        //                 } else {
-	        //                     hBeta[j] = 0.0;
-	        //                 }
-	        //             }
-
-	        // xBetaKnown = false;
-	        // checkAllLazyFlags();
-	        modelSpecifics.computeXBeta(hBeta.data(), useCrossValidation);
-
-
-	        //             done = check();
-	        //             if (done) {
-	        //                 std::cerr << "break B" << std::endl;
-	        //                 break;
-	        //             }
-
-	        double qnObjective = getLogLikelihood() + getLogPrior();
-
-	        if (ccdObjective > qnObjective) { // Revert
-	            Map<VectorXd>(hBeta.data(), J) = x; // Set CCD solution
-	            modelSpecifics.computeXBeta(hBeta.data(), useCrossValidation);
-	            // xBetaKnown = false;
-	            // checkAllLazyFlags();
-
-	            double ccd2Objective = getLogLikelihood() + getLogPrior();
-
-	            if (ccdObjective != ccd2Objective) {
-	                std::cerr << "Poor revert: " << ccdObjective << " != " << ccd2Objective << " diff: "<< (ccdObjective - ccd2Objective) << std::endl;
-	            }
-	            // lastObjFunc = savedLastObjFunc;
-	            std::cerr << "revert" << std::endl;
-	        } else {
-	            std::cerr << "accept" << std::endl;
-	            //                 done = check();
-	            //                 if (done) {
-	            //                     std::cerr << "break C" << std::endl;
-	            //                     break;
-	            //                 }
-	        }
-
-	        done = check();
-
-	        // lastObjFunc = getLogLikelihood() + getLogPrior();
-
-	        // Get ready for next secant-pair
-	        previousSecant = newestSecant;
-	        newestSecant = (newestSecant + 1) % qnQ;
-	    }
+	    // using namespace Eigen;
+	    //
+	    // Eigen::MatrixXd secantsU(J, qnQ);
+	    // Eigen::MatrixXd secantsV(J, qnQ);
+	    //
+	    // VectorXd x(J);
+	    //
+	    // // Fill initial secants
+	    // int countU = 0;
+	    // int countV = 0;
+	    //
+	    // for (int q = 0; q < qnQ; ++q) {
+	    //     x = Map<const VectorXd>(hBeta.data(), J); // Make copy
+	    //
+	    //     cycle();
+	    //     done = check();
+	    //     if (done) break;
+	    //
+	    //     if (countU == 0) { // First time through
+	    //         secantsU.col(countU) = Map<const VectorXd>(hBeta.data(), J) - x;
+	    //         ++countU;
+	    //     } else if (countU < qnQ - 1) { // Middle case
+	    //         secantsU.col(countU) = Map<const VectorXd>(hBeta.data(), J) - x;
+	    //         secantsV.col(countV) = secantsU.col(countU);
+	    //         ++countU;
+	    //         ++countV;
+	    //     } else { // Last time through
+	    //         secantsV.col(countV) = Map<const VectorXd>(hBeta.data(), J) - x;
+	    //         ++countV;
+	    //     }
+	    // }
+	    //
+	    // // 	    std::cerr << secantsU << std::endl << std::endl;
+	    // // 	    std::cerr << secantsV << std::endl;
+	    //
+	    // int newestSecant = qnQ - 1;
+	    // int previousSecant = newestSecant - 1;
+	    //
+	    // while (!done) {
+	    //
+	    //     // 2 cycles for each QN step
+	    //     x = Map<const VectorXd>(hBeta.data(), J); // Make copy
+	    //     cycle();
+	    //     // 	        done = check();
+	    //     // 	        if (done) {
+	    //     // 	            std::cerr << "break A" << std::endl;
+	    //     // 	            break;
+	    //     // 	        }
+	    //
+	    //     secantsU.col(newestSecant) = Map<const VectorXd>(hBeta.data(), J) - x;
+	    //
+	    //     VectorXd Fx = Map<const VectorXd>(hBeta.data(), J); // TODO Can remove?
+	    //
+	    //     x = Map<const VectorXd>(hBeta.data(), J);
+	    //     cycle();
+	    //     // 	        done = check();
+	    //     // 	        if (done) {
+	    //     // 	            std::cerr << "break B" << std::endl;
+	    //     // 	            break;
+	    //     // 	        }
+	    //
+	    //     secantsV.col(newestSecant) = Map<const VectorXd>(hBeta.data(), J) - x;
+	    //
+	    //     // Do QN step here
+	    //
+	    //     //             MatrixXd UtU = secantsU.transpose() * secantsU;
+	    //     //             MatrixXd UtV = secantsU.transpose() * secantsV;
+	    //     //             MatrixXd M = UtU - UtV;
+	    //
+	    //     auto M = secantsU.transpose() * (secantsU - secantsV);
+	    //
+	    //     auto Minv = M.inverse();
+	    //
+	    //     auto A = secantsU.transpose() * secantsU.col(newestSecant);
+	    //     auto B = Minv * A;
+	    //     auto C = secantsV * B;
+	    //
+	    //     // MatrixXd A = secantsV * Minv;
+	    //     // MatrixXd B = A * secantsU.transpose();
+	    //     // MatrixXd C = B * secantsV.col(newestSecant); // - (x - Fx)
+	    //     // MatrixXd C = B * secantsU.col(newestSecant); // TODO Can remove?
+	    //
+	    //     //     VectorXd xqn = Map<const VectorXd>(hBeta.data(), J) + C;
+	    //     VectorXd xqn = Fx + C; // TODO Can remove?
+	    //
+	    //     // Save CCD solution
+	    //     x = Map<const VectorXd>(hBeta.data(), J);
+	    //     double ccdObjective = getLogLikelihood() + getLogPrior();
+	    //     // double savedLastObjFunc = lastObjFunc;
+	    //
+	    //     Map<VectorXd>(hBeta.data(), J) = xqn; // Set QN solution
+	    //     //             for (int j = 0; j < J; ++j) {
+	    //     //                 if (sign(hBeta[j]) == xqn(j)) {
+	    //     //                     hBeta[j] = xqn(j);
+	    //     //                 } else {
+	    //     //                     hBeta[j] = 0.0;
+	    //     //                 }
+	    //     //             }
+	    //
+	    //     // xBetaKnown = false;
+	    //     // checkAllLazyFlags();
+	    //     modelSpecifics.computeXBeta(hBeta.data(), useCrossValidation);
+	    //
+	    //
+	    //     //             done = check();
+	    //     //             if (done) {
+	    //     //                 std::cerr << "break B" << std::endl;
+	    //     //                 break;
+	    //     //             }
+	    //
+	    //     double qnObjective = getLogLikelihood() + getLogPrior();
+	    //
+	    //     if (ccdObjective > qnObjective) { // Revert
+	    //         Map<VectorXd>(hBeta.data(), J) = x; // Set CCD solution
+	    //         modelSpecifics.computeXBeta(hBeta.data(), useCrossValidation);
+	    //         // xBetaKnown = false;
+	    //         // checkAllLazyFlags();
+	    //
+	    //         double ccd2Objective = getLogLikelihood() + getLogPrior();
+	    //
+	    //         if (ccdObjective != ccd2Objective) {
+	    //             std::cerr << "Poor revert: " << ccdObjective << " != " << ccd2Objective << " diff: "<< (ccdObjective - ccd2Objective) << std::endl;
+	    //         }
+	    //         // lastObjFunc = savedLastObjFunc;
+	    //         std::cerr << "revert" << std::endl;
+	    //     } else {
+	    //         std::cerr << "accept" << std::endl;
+	    //         //                 done = check();
+	    //         //                 if (done) {
+	    //         //                     std::cerr << "break C" << std::endl;
+	    //         //                     break;
+	    //         //                 }
+	    //     }
+	    //
+	    //     done = check();
+	    //
+	    //     // lastObjFunc = getLogLikelihood() + getLogPrior();
+	    //
+	    //     // Get ready for next secant-pair
+	    //     previousSecant = newestSecant;
+	    //     newestSecant = (newestSecant + 1) % qnQ;
+	    // }
     } else { // No QN
         while (!done) {
             cycle();
@@ -1246,6 +1259,18 @@ double CyclicCoordinateDescent::getHessianDiagonal(int index) {
 	return g_d2;
 }
 
+double CyclicCoordinateDescent::getJerkDiagonal(int index) {
+
+    checkAllLazyFlags();
+
+    double jerk;
+
+    computeNumeratorForGradient(index); // TODO Is this necessary?
+    modelSpecifics.computeThirdDerivative(index, &jerk, useCrossValidation);
+
+    return jerk;
+}
+
 double CyclicCoordinateDescent::getAsymptoticVariance(int indexOne, int indexTwo) {
 	checkAllLazyFlags();
 	if (!fisherInformationKnown) {
@@ -1268,22 +1293,22 @@ double CyclicCoordinateDescent::getAsymptoticVariance(int indexOne, int indexTwo
 	}
 }
 
-double CyclicCoordinateDescent::getAsymptoticPrecision(int indexOne, int indexTwo) {
-	checkAllLazyFlags();
-	if (!fisherInformationKnown) {
-		computeAsymptoticPrecisionMatrix();
-		fisherInformationKnown = true;
-	}
-
-	IndexMap::iterator itOne = hessianIndexMap.find(indexOne);
-	IndexMap::iterator itTwo = hessianIndexMap.find(indexTwo);
-
-	if (itOne == hessianIndexMap.end() || itTwo == hessianIndexMap.end()) {
-		return NAN;
-	} else {
-		return hessianMatrix(itOne->second, itTwo->second);
-	}
-}
+// double CyclicCoordinateDescent::getAsymptoticPrecision(int indexOne, int indexTwo) {
+// 	checkAllLazyFlags();
+// 	if (!fisherInformationKnown) {
+// 		computeAsymptoticPrecisionMatrix();
+// 		fisherInformationKnown = true;
+// 	}
+//
+// 	IndexMap::iterator itOne = hessianIndexMap.find(indexOne);
+// 	IndexMap::iterator itTwo = hessianIndexMap.find(indexTwo);
+//
+// 	if (itOne == hessianIndexMap.end() || itTwo == hessianIndexMap.end()) {
+// 		return NAN;
+// 	} else {
+// 		return hessianMatrix(itOne->second, itTwo->second);
+// 	}
+// }
 
 CyclicCoordinateDescent::Matrix CyclicCoordinateDescent::computeFisherInformation(const std::vector<IdType>& indices) const {
     Matrix fisherInformation(indices.size(), indices.size());
@@ -1382,7 +1407,7 @@ void CyclicCoordinateDescent::mmUpdateAllBeta(std::vector<double>& delta,
 
             gh[j].second /= scale;
 
-            delta[j] = jointPrior->getDelta(gh[j], hBeta, j);
+            delta[j] = jointPrior->getDelta(gh[j], hBeta, j, *this);
             // TODO this is only correct when joints are independent across dimensions
         } else {
             delta[j] = 0.0;
@@ -1409,7 +1434,7 @@ double CyclicCoordinateDescent::ccdUpdateBeta(int index) {
 	    gh.second = 0.0;
 	}
 
-	return jointPrior->getDelta(gh, hBeta, index);
+	return jointPrior->getDelta(gh, hBeta, index, *this);
 }
 
 void CyclicCoordinateDescent::axpyXBeta(const double beta, const int j) {